Create a new renderer for UI
Old string renderer will be removed eventually in favor of having a UI renderer which can handle strings and images.
This will also be able to render a string at any font, typeface, and size eventually.
Doesnt actually do anything yet, as vertex and index buffers still need to be made and managed.