Update log system to use 2 log files
The first log file (active.log) is always in the root directory of the current working directory and is named the same for each runtime execution. This can be monitored by external programs or by the user to see the log for the current active process or the last run of the process.
The second log file (log/<timestamp>.log) is the archive of the last run of the process. It is also written at live, but the name of the file is based on the timestamp of when the process was launched.